requirements.txt
python-dotenv
africastalking
openai
requests
folium
geopy
langchain
langchain-community
langchain-core
langchain-google-genai
google-generativeai
python-dotenv
chromadb
faiss-cpu
pypdf
google-genai
plotly
pandas
